refactored all the admin files to a new admin dir for cleaner, more findable code
